feat/trigger release from gh action (#819)
This PR introduces Argo sensors (and eventbus and a few other related services) together with dedicated GitHub Actions to facilitate the data release process. In a nutshell, when someone triggers a data-release pipeline (either through the Kedro pipeline "kg_release" or through "data_release"), it will –on success– create a PR that contains a draft of the release notes and an associated article, together with parameters that should be published to the Every Cure website, so people may easily find out which parameters (a subset thereof) were associated with a certain data release.

---------

Co-authored-by: emil-k <emil.krause.44@gmail.com>
Co-authored-by: Laurens Vijnck <laurens@everycure.org>
Co-authored-by: Pascal Brokmeier <pascal@everycure.org>
Co-authored-by: Piotr Kaniewski <115791652+piotrkan@users.noreply.github.com>
Co-authored-by: leelancashire <drllancashire@gmail.com>
Co-authored-by: Mateusz <39764611+matwasilewski@users.noreply.github.com>
Co-authored-by: Laurens <90421718+lvijnck@users.noreply.github.com>
Co-authored-by: Alexei Stepanenko <alexei.stepa@gmail.com>
Co-authored-by: Siyan Luo <siyanluo@Siyans-MacBook-Pro.local>
Co-authored-by: All of us at Every Cure <releasebot@everycure.org>