feat: allow intermediate non-latest releases (#957)
* feat: allow intermediate non-latest releases

These shouldn't be labeled as the latest release in GitHub.

* fix: add basic metadata on the release notes and post of 0.2.3

v0.2.3 is an intermediate release, triggered to evaluate the good working of the
data release workflow. An oversight lead into the generation of these nodes and posts
without having the necessary frontmatter for mkdocs, which is triggered from a
workflow that eluded us.