refactor: change mock test to integration
In the previous version, all that got checked was that  was called at
least once.  For a test that checks whether or not the function can
interact with an external service, that's unuseful.  This commit
modifies the test and the corresponding code to execute only the
commands that are really needed, without changing settings in a
trigger-happy fashion.