Added to pipeline registry
Add test_no_modelling to pipeline registry for a faster test for the qc metric logging. Modified make file to run this with make integration_test.

This removes the embeddings and modeling from the integration_test.

Comments out the mlflow hook flags in .env.defaults so that mlflow is on and logging locally