Render the github.event content in CI without interpolation, avoiding prematurely closing it (#1107)
* fix: prevent shell from interpreting tokens in string

It was interpreting e.g. the GH PR number '#912' as a comment, thus not recognizing the closing quotes that came later on that line.

* fix: allow SA key to be deleted by user

* fix: use a heredoc without interpolation of the string

The next issue that was encountered when using the previous ... is that
single quotes could be present in the placeholder. Since EOM is a fairly
common heredoc delimiter, I'll use something recognizable, but far less
common.

* added comment to explain the rather uncommon here-doc

---------

Co-authored-by: Laurens <90421718+lvijnck@users.noreply.github.com>