fix: ensure batches won't get skipped
If the API gets called, it needs to be passed a stable slice of an iterable.
If a retry-mechanism is added, an iterable means it will have consumed a part
of the iterable in a call that errored out, so rather than processing 100%
of the iterable, less than that would've been processed.

This is easy to fix, by wrapping only the call to the API, not including the iterable consumption.