allow for better concurrency
the semaphore is only there to 'protect' the API, not the writing to GCP or local disk. By introducing the asyncio.sleep, there are better chances of being able to trigger another embedding call in a different thread (cooperative multitasking). In the meantime, one could also write away the data to GCS.