add workaround for quasi all null values to Spark
It turns out that SPark cannot properly infer the schema of a column consisting of all nulls when the file containing that column has not been written by Spark but by e.g. pandas. To fix that, one would need to add metadata to the schema, which we have little control over.