Reduce the likelihood of creating null-only columns for which Spark can't infer the datatype without metadata.   (#1362)
* Revert "fix: drop column from data that is manually lifted to Hive partition layout"

This reverts commit 2e02a36d65bb5f8fd57bf5b61d90580fae3e5f9e.

* fix: drop column from data that is manually lifted to Hive partition layout

* add workaround for quasi all null values to Spark

It turns out that SPark cannot properly infer the schema of a column consisting of all nulls when the file containing that column has not been written by Spark but by e.g. pandas. To fix that, one would need to add metadata to the schema, which we have little control over.

* fix: with lower logging, don't sort nulls against strings