Adding the legacy LLM tagging pipeline back into the disease list
In the long term we should get rid of this step, but for now, we are pulling the LLM generated tags from the old disease list repo into the pipeline to make sure we still have these columns present. These columns need some good amount of cleaning, adding to the verbosity of this commit.