Rename `integration.prm.unified_nodes` (and edges) datasets to `integration.prm.unified_nodes@spark` (#1982)
* Rename integration.prm.unified_nodes to integration.prm.unified_nodes@spark

This is to be more aligned with the convention

* Update dataset references to use Spark variant

Changed references from 'integration.prm.unified_nodes' to 'integration.prm.unified_nodes@spark' in documentation in line with the renaming of the dataset.

* Add kedro catalog resolve for sample environment

Introduces a new CI step to run 'kedro catalog resolve' with the 'sample' environment, ensuring catalog resolution is tested for this configuration.

* Update catalog source names for Spark integration in sample pipeline

Renamed 'integration.prm.unified_nodes' and 'integration.prm.unified_edges' in sample pipeline to include '@spark' suffix.